added position service, dao, mapper, controller, model.

+++ src/main/java/com/controlcenter/controlcenter/model/Position.java
@@ -11,7 +11,7 @@ public class Position {
     private Long position_id;
     private String position_name;
     private String position_sh_name;
-    private Long del_flag;
+    private int del_flag;
     private String reg_id;
     private Timestamp reg_date;
     private String update_id;
