+++ src/main/java/com/controlcenter/controlcenter/model/User.java
@@ -18,7 +18,7 @@ public class User {
   private String status_code;
   private Long role_id;
   private String img_src;
-  private Long del_flag;
+  private int del_flag;
   private String reg_id;
   private Timestamp reg_date;
   private String update_id;
